This adds two improvements:
     * Change the computation of the `_size` attribute for expressions, to be more
       sensitive to some of the integer coefficients, e.g., `4*a` is now structurally
       larger than `3*a`. Similarly with `a**4` and `a**3`. This allows the
       `_syntactic_cmp` to short circuit most of the comparison. This operation
       is used a lot when sorting and normalizing the representation of symbolic
       expressions.
      * Change the caching of the bounds computation. Now we store in the cache
       the precision with which we computed the previous bounds, which allows
       better reuse of the cache.

 On a microbenchmark, this resulted in a reduction by 30% (before 3210 and after 2145)
 of the number of calls to `bounds_for_sorted_terms` and a 10% reduction in the
 total time spent in `bounds`:

 After:
```
    tests/shape_poly_test.py::ShapePolyTest::test_constraints_for_profile          2307348 function calls (2260293 primitive calls) in 0.962 seconds
       1    0.000    0.000    0.969    0.969 shape_poly_test.py:1580(test_constraints_for_profile)
       1    0.000    0.000    0.234    0.234 shape_poly_test.py:1583(f)
     320    0.000    0.000    0.095    0.000 _shape_poly_decision.py:41(bounds_decision)
 425/280    0.001    0.000    0.094    0.000 _shape_poly_decision.py:234(bounds)
  513/51    0.002    0.000    0.091    0.002 _shape_poly_decision.py:260(_bounds_for_sorted_terms)
1230/135    0.001    0.000    0.081    0.001 _shape_poly_decision.py:330(add_implicit_constraints)
     250    0.000    0.000    0.076    0.000 _shape_poly.py:784(__ge__)
     250    0.000    0.000    0.076    0.000 _shape_poly.py:1077(_geq_decision)
 381/289    0.001    0.000    0.069    0.000 _shape_poly_decision.py:102(combine_and_add_constraint)
     695    0.001    0.000    0.065    0.000 _shape_poly.py:1673(_evaluate_multiply)
3572/766    0.002    0.000    0.051    0.000 _shape_poly.py:637(__str__)
```

 Before:
```
tests/shape_poly_test.py::ShapePolyTest::test_constraints_for_profile          3486289 function calls (3318484 primitive calls) in 1.240 seconds
       1    0.000    0.000    1.247    1.247 shape_poly_test.py:1569(test_constraints_for_profile)
 992/320    0.001    0.000    0.424    0.001 _shape_poly_decision.py:269(bounds)
3210/280    0.008    0.000    0.423    0.002 _shape_poly_decision.py:292(_bounds_for_sorted_terms)
     250    0.000    0.000    0.400    0.002 _shape_poly.py:783(__ge__)
     250    0.000    0.000    0.399    0.002 _shape_poly_decision.py:39(geq_decision)
```